+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `global_seq` and `group_seq` to `Client`, sequence numbers assigned to test instance by the sync service. See [PR 29]
 
 ### Change
+- Move `RunParameters` to a field on `Client`. See [PR 29].
+
+- Don't wait for network when no sidecar. See [PR 27].
+
 - Make `RunParameters::test_group_instance_count` a `u64` to be consistent with
   `RunParameters::test_instance_count`. See [PR 26].
 
@@ -17,10 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   other instances to do the same. Also makes `Client::wait_network_initialized`
   private, as it is included in `Client::new_and_init` now. See [PR 25].
 
-- Move `RunParameters` to a field on `Client`. See [PR 29].
-
 [PR 26]: https://github.com/testground/sdk-rust/pull/26
 [PR 26]: https://github.com/testground/sdk-rust/pull/25
+[PR 27]: https://github.com/testground/sdk-rust/pull/27
 [PR 29]: https://github.com/testground/sdk-rust/pull/29
 
 ## [0.3.0]