+++ CHANGELOG.md
@@ -22,7 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   private, as it is included in `Client::new_and_init` now. See [PR 25].
 
 [PR 26]: https://github.com/testground/sdk-rust/pull/26
-[PR 26]: https://github.com/testground/sdk-rust/pull/25
+[PR 25]: https://github.com/testground/sdk-rust/pull/25
 [PR 27]: https://github.com/testground/sdk-rust/pull/27
 [PR 29]: https://github.com/testground/sdk-rust/pull/29