+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - unreleased
+### Change
+- Change `RunParameters::test_start_time` from `String` to `DateTime` for ease of use. See [PR 41].
+
+[PR 41]: https://github.com/testground/sdk-rust/pull/41
+
 ## [0.4.0]
 ### Added
 - Add `global_seq` and `group_seq` to `Client`, sequence numbers assigned to test instance by the sync service. See [PR 29]